Cut Runtime Cache reads and writes for WCL data
The Hobby Runtime Cache is shared across all projects in the team and has
monthly read/write caps; this project was the dominant consumer. Three
changes reduce its footprint without fetching from WCL any more often:

- Merge zones, classes and regions into a single `getGameData` remote cache
  entry, fetched with one combined GraphQL query. Every render and every
  filter picker now reads one Runtime Cache entry instead of three, and a
  miss makes one WCL request instead of three.

- Lengthen the rankings cacheLife from 1h revalidate / 30d expire to 24h,
  with `revalidate === expire` to disable stale-while-revalidate. A rankings
  key is now rewritten at most once a day instead of on every access after it
  goes stale, which lifts the cache hit rate and cuts writes.

- Move the WCL OAuth token from `use cache: remote` to in-memory `use cache`.
  The token is valid ~360 days, so per-instance caching is enough and keeps
  it off the Runtime Cache meter entirely.
